refactor auth mechanism into plugin
