+++ plugins/loader.js
@@ -1,0 +1,3 @@
+exports.load = function(pluginType, pluginName) {
+  return require('./' + pluginType + '/' + pluginName + '/' + pluginName);
+};
